Mostly commenting and moving stuff about
Changed the order and added comments to explain what things are - and a
bit on why. Trying to get the multivariate normal example to look like
my target in the example schema - I still don't like the array types as
they are defined - too clumsy with multiple value tags - covariance
matrix is worse!